Integrated django-allauth
In order to provide user management the external app django-allauth is
used. It is set up so that a user has to provide an email address for
verification on signup. There is no social login integration enabled at
the moment. The verification emails are sent to the console.
To provide the structure to override the allauth templates a
corresponding directory is created and referenced in the settings, as
well as an overridden login.html for proof of concept.
Additionally, a very basic infrastructure for templates in the entire
application is set up.