Enhance README and add drafts directory
The README got more important sections, namely Philosophy, Get in touch
and Contributing. It now contains information to contact the developers
and getting ideas about contributing to the project. The README still
misses a section about the actual content of the project, for now it
just links to the drafts.
The drafts directory contains some ideas for the platform in the form of
a mind map and a rough domain model that will be adjusted over the time
of development to fit the needs, but it gives a starting point. This is
also the place where UI sketches will be located as soon as there are
some available.