Custom user model and proper profile page
In order to be able to store custom properties for users the default
user model of django is extended. For proof of concept the field
total_amount_of_kilometers is added.
Additionally, the profile page was updated. It is now accessible via
the menu bar for authenticated users. If someone who is not
authenticated tries to access the page, he/she is redirected to the
login page and, on successful authentication, redirected again to
the profile page.