Management of Rides for users
If a user is logged in there is a possibility to create a new ride
(link in the navbar). A ride consists of date, start time, end
time, distance in kilometers and a note.
On the profile page (only visible for logged users as well) there
is a table displaying the rides of the user and a sum of kilometers
that the user has completed.