Bugfix in ProfileView and documentation
Before this commit the profile view showed all rides that were in
the DB, but did not filter for the current user. Also, the
implementation of the summation of distances was quite inefficient,
this was improved as well by reducing DB access to one query.
In addition, some documentation in the views is provided.