Bootstrap include and navbar styling
This commit contains the initial bootstrap integration. For now,
it is used to create and style the navbar, which is different for
logged and non-logged users, obviously. The content of the pages has
not been changed.
There are some fake options in the navbar as well which will either
be populated or removed or changed in the future.
The old "navbar" (collection of working links on the top of each
page) has been removed.
The style change does not affect pages which do not extend _base,
which are at the moment most of the pages that are provided by
django-allauth.