CSS Refactoring and simple footer
The existing CSS file was renamed to custom.css in order to be able
to put all custom styles in.
The base template is being extended by a footer bar, containing
the project name, links to about and data page (both not existing
yet) and a github link.
This commit also includes FontAwesome integration.