Signup view styling
Using the base layout of the login screen, the signup screen was
updated with style and layout. It consists of the fields email,
username, password1 and password2.
Currently, there is no mechanism to tell the user that signup was
not successful (e.g. username and/or email already exist).
Also, the signup page redirects to the email verification page,
which has no styling yet.