Password Change Page Styling
The template for password change is being replaced by a custom one which
uses the application styling. The user has to enter his/her current
password for verification and then the new one twice.
At the moment, there is no user feedback about success or failure, and
the django-allauth view handler does not support redirect. This has to
be added in the future, as well as message generation on failure
(allauth only emits success messages, failures are not explicitly sent).
In order to be able to use the password change function, the menu bar
was extended as the settings page does not exist yet.