Email Settings Page style
This commit contains the new email settings page. It displays all email
addresses of the user in a table and provides options to re-send the
verification code, make an address primary or delete it. For this to
work the EmailView of django-allauth is extended.
Additionally, as on the original page, the user is able to add new
addresses.