Profile view re-design
The profile view now shows a table of the 10 most recent rides of
the user. There are links to view all all rides (not working yet)
and to add a ride (redirects to the not-finished page for entering
rides).
Additionally, there is a table which will show the top routes of
the user, but as routes are not implemented yet, this is the
identical table as above, showing the top 10 rides of the user.
There are links for a complete route list and route management as
well, both of which are not implemented yet.