Entering and saving routes
This commit introduces the route model class. A route is a template
for a ride, so the user does not have to enter distance and start
or destination for rides he/she does more than once. Routes have
to have a name and a distance, but start, destination and note are
optional.
The routes of the user are displayed on the profile page in the
second table.
A user can enter routes with a corresponding form, reachable via
the navbar.
Currently, the name of a route is marked as unique for the database,
independent of the user. So currently it is not possible for two
different users to have a route with the same name.
As always, there is no usable feedback for the user upon completion
of a created route. The user is only re-directed, regardless of
successful operation or any error.