Entering Rides Take II
Entering rides is now possible, including a re-styled UI and the
option to either enter a distance directly or reference a route in
the ride.
The fields start_time, end_time and note of the ride have become
optional and may now be NULL in the database.
The profile view has seen a little bit of fine tuning, so the rides
table respects rides with and without routes, and automatically
displays the distance from the route if the ride references a route.
Additionally, the note field in the rides table shows the name of
the selected route if there is one.