Styling of password reset screens
When the user forgot his/her password and wants to reset it, he/she
may enter an email on a separate page (the password reset page).
After that, he/she is redirected to the password reset done page.

Both of the templates have been replaced from django-allauth to
match the style of the application and include the correct navbar.